Updated gonum version
gonum has moved to a new Go package repository. This commits updates all
the package imports to gonum.org/v1/gonum/mat and changes few source
files to to this change.

Another thing this commit brings is dependency management. I decided to
use go dep tool which has become a part of go toolchain and having had a
positive experience with it I chose to use it for this project.